Reorganize and rename head files
