+++ packages/inertia/src/head.js
