+++ packages/inertia/src/index.js
@@ -1,3 +1,4 @@
 export { default as Inertia } from './inertia'
+export { default as createHeadManager } from './head'
 export { default as shouldIntercept } from './shouldIntercept'
 export { hrefToUrl, mergeDataIntoQueryString, urlWithoutHash } from './url'
