+++ packages/inertia/src/inertia.js
@@ -5,7 +5,6 @@ import { fireBeforeEvent, fireErrorEvent, fireExceptionEvent, fireFinishEvent, f
 import { hrefToUrl, mergeDataIntoQueryString, urlWithoutHash } from './url'
 import { hasFiles } from './files'
 import { objectToFormData } from './formData'
-import createMetaManager from './meta'
 
 export default {
   resolveComponent: null,
@@ -23,12 +22,6 @@ export default {
     this.transformProps = transformProps || this.transformProps
     this.handleInitialPageVisit(initialPage)
     this.setupEventListeners()
-  },
-
-  initInstance(isServer) {
-    return {
-      meta: createMetaManager(isServer),
-    }
   },
 
   handleInitialPageVisit(page) {
@@ -402,13 +395,15 @@ export default {
   },
 
   remember(data, key = 'default') {
-    this.replaceState({
-      ...this.page,
-      rememberedState: {
-        ...this.page.rememberedState,
-        [key]: data,
-      },
-    })
+    if (this.page) {
+      this.replaceState({
+        ...this.page,
+        rememberedState: {
+          ...this.page.rememberedState,
+          [key]: data,
+        },
+      })
+    }
   },
 
   restore(key = 'default') {
